requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.3.2
decorator==4.4.2
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
factory-boy==3.3.0
Faker==19.10.0
idna==3.4
imageio==2.31.5
imageio-ffmpeg==0.4.9
inflection==0.5.1
iniconfig==2.0.0
kombu==5.3.2
moviepy==1.0.3
numpy==1.26.0
packaging==23.2
Pillow==10.0.1
pluggy==1.3.0
proglog==0.1.10
prompt-toolkit==3.0.39
pytest==7.4.2
pytest-django==4.5.2
pytest-factoryboy==2.6.0
python-dateutil==2.8.2
pytz==2023.3.post1
redis==4.6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
vine==5.0.0
wcwidth==0.2.8
